src/test/java/ru/netology/stats/StatsServiceTest.java
package ru.netology.stats;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class StatsServiceTest {

    @org.junit.jupiter.api.Test
    void shouldTestCalculateTotal() {
        StatsService service = new StatsService();
        int [] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 180;
        int actual = service.calculateTotal(sales);

        assertEquals(expected,actual);
    }


    @org.junit.jupiter.api.Test
    void shouldTestCalculateAverage() {
        StatsService service = new StatsService();
        int [] monthlySales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 15;
        int actual = service.calculateAverage(monthlySales);

        assertEquals(expected,actual);
    }

    @Test
    void shouldTestFindMonthMax() {
        StatsService service = new StatsService();
        int [] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 8;
        int actual = service.findMonthMax(sales);

        assertEquals(expected,actual);
    }

    @Test
    void shouldTestFindMin() {
        StatsService service = new StatsService();
        int [] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 9;
        int actual = service.findMonthMin(sales);

        assertEquals(expected, actual);
    }


    @Test
    void shouldTestCalculateMonthBelowAverage() {
        StatsService service = new StatsService();
        int [] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 5;
        int actual = service.calculateMonthBelowAverage(sales);

        assertEquals(expected, actual);
    }


    @Test
    void shouldTestCalculateMonthAboveAverage() {
        StatsService service = new StatsService();
        int [] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 5;
        int actual = service.calculateMonthAboveAverage(sales);

        assertEquals(expected, actual);
    }
}